更新用例快照，增强稳定性
# message auto-generated for no-merge-commit merge:
merge qcc-update-snapshots into master

更新 GUI 用例快照，增强用例稳定性

Created-by: insightDeveloper

Match-id-0252d21c473987654eccaec12f5221ffa756414c

+++ framework/e2e/utils/index.ts
@@ -45,12 +45,16 @@ export async function importData(page: Page, filePath: FilePath = FilePath.TEXT)
 export async function clearAllData(page: Page): Promise<void> {
     const frameworkPage = new FrameworkPage(page);
     const { settingsBtn, deleteAllBtn, deleteAllDialog, deleteAllConfirmBtn, projectList } = frameworkPage;
+    const isSettingsBtnDisabled = await settingsBtn.evaluate((el) => el.classList.contains('disabled'));
 
+    if (isSettingsBtnDisabled) {
+        return;
+    }
     await settingsBtn.click();
 
-    const isDisabled = await deleteAllBtn.evaluate((el) => el.classList.contains('disabled'));
+    const isDeleteBtnDisabled = await deleteAllBtn.evaluate((el) => el.classList.contains('disabled'));
 
-    if (isDisabled) {
+    if (isDeleteBtnDisabled) {
         return;
     }
     await deleteAllBtn.click();
